// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  CUSTOMER: 'CUSTOMER',
  MERCHANT: 'MERCHANT',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const QueueStatus = {
  OPEN: 'OPEN',
  PAUSED: 'PAUSED',
  CLOSED: 'CLOSED'
} as const

export type QueueStatus = (typeof QueueStatus)[keyof typeof QueueStatus]


export const QueueEntrySource = {
  ONLINE: 'ONLINE',
  WALK_IN: 'WALK_IN'
} as const

export type QueueEntrySource = (typeof QueueEntrySource)[keyof typeof QueueEntrySource]


export const QueueEntryStatus = {
  WAITING: 'WAITING',
  CHECKED_IN: 'CHECKED_IN',
  CALLED: 'CALLED',
  SERVING: 'SERVING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW'
} as const

export type QueueEntryStatus = (typeof QueueEntryStatus)[keyof typeof QueueEntryStatus]


export const NotificationType = {
  QUEUE_GETTING_CLOSE: 'QUEUE_GETTING_CLOSE',
  CUSTOMER_CALLED: 'CUSTOMER_CALLED',
  QUEUE_UPDATED: 'QUEUE_UPDATED'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]